+++ packages/gatsby/src/bootstrap/requires-writer.js
@@ -2,6 +2,8 @@ const _ = require(`lodash`)
 const fs = require(`fs-extra`)
 const crypto = require(`crypto`)
 const { store, emitter } = require(`../redux/`)
+const reporter = require(`gatsby-cli/lib/reporter`)
+const { match } = require(`@reach/router/lib/utils`)
 import { joinPath } from "gatsby-core-utils"
 
 let lastHash = null
@@ -18,23 +20,59 @@ const getComponents = pages =>
     .map(pickComponentFields)
     .uniqBy(c => c.componentChunkName)
     .value()
+
 /**
  * Get all dynamic routes and sort them by most specific at the top
  * code is based on @reach/router match utility (https://github.com/reach/router/blob/152aff2352bc62cefc932e1b536de9efde6b64a5/src/lib/utils.js#L224-L254)
  */
 const getMatchPaths = pages => {
-  const filteredPaths = []
+  const createMatchPathEntry = (page, index) => {
+    let score = page.matchPath.replace(/\/$/, ``).split(`/`).length
+    if (!page.matchPath.includes(`*`)) {
+      score += 1
+    }
+
+    return {
+      ...page,
+      index,
+      score,
+    }
+  }
+
+  const matchPathPages = []
   pages.forEach((page, index) => {
     if (page.matchPath) {
-      filteredPaths.push({
-        ...page,
-        index,
-        score: page.matchPath.split(`/`).length,
-      })
+      matchPathPages.push(createMatchPathEntry(page, index))
     }
   })
 
-  return filteredPaths
+  // Pages can live in matchPaths, to keep them working without doing another network request
+  // we save them in matchPath. Our sorting will put them above dynamic routes
+  // as we add a static bonus point to static routes
+  // More info in https://github.com/gatsbyjs/gatsby/issues/16097
+  // small speedup: don't bother traversing when no matchPaths found.
+  if (matchPathPages.length) {
+    pages.forEach((page, index) => {
+      const isInsideMatchPath = !!matchPathPages.find(
+        pageWithMatchPath =>
+          !page.matchPath && match(pageWithMatchPath.matchPath, page.path)
+      )
+
+      if (isInsideMatchPath) {
+        matchPathPages.push(
+          createMatchPathEntry(
+            {
+              ...page,
+              matchPath: page.path,
+            },
+            index
+          )
+        )
+      }
+    })
+  }
+
+  return matchPathPages
     .sort((a, b) => {
       // The higher the score, the higher the specificity of our matchPath
       const order = b.score - a.score
@@ -42,7 +80,7 @@ const getMatchPaths = pages => {
         return order
       }
 
-      // if specificty is the same we use the array index
+      // if specificity is the same we use the array index
       return b.index - a.index
     })
     .map(({ path, matchPath }) => {
@@ -58,6 +96,7 @@ const createHash = (matchPaths, components) =>
 
 // Write out pages information.
 const writeAll = async state => {
+  // console.log(`on requiresWriter progress`)
   const { program } = state
   const pages = [...state.pages.values()]
   const matchPaths = getMatchPaths(pages)
@@ -67,7 +106,8 @@ const writeAll = async state => {
 
   if (newHash === lastHash) {
     // Nothing changed. No need to rewrite files
-    return Promise.resolve()
+    // console.log(`on requiresWriter END1`)
+    return false
   }
 
   lastHash = newHash
@@ -110,18 +150,34 @@ const preferDefault = m => m && m.default || m
       .then(() => fs.move(tmp, destination, { overwrite: true }))
   }
 
-  const result = await Promise.all([
+  await Promise.all([
     writeAndMove(`sync-requires.js`, syncRequires),
     writeAndMove(`async-requires.js`, asyncRequires),
     writeAndMove(`match-paths.json`, JSON.stringify(matchPaths, null, 4)),
   ])
 
-  return result
+  return true
 }
 
-const debouncedWriteAll = _.debounce(() => writeAll(store.getState()), 500, {
-  leading: true,
-})
+const debouncedWriteAll = _.debounce(
+  async () => {
+    const activity = reporter.activityTimer(`write out requires`, {
+      id: `requires-writer`,
+    })
+    activity.start()
+    const didRequiresChange = await writeAll(store.getState())
+    if (didRequiresChange) {
+      reporter.pendingActivity({ id: `webpack-develop` })
+    }
+    activity.end()
+  },
+  500,
+  {
+    // using "leading" can cause double `writeAll` call - particularly
+    // when refreshing data using `/__refresh` hook.
+    leading: false,
+  }
+)
 
 /**
  * Start listening to CREATE/DELETE_PAGE events so we can rewrite
@@ -129,18 +185,22 @@ const debouncedWriteAll = _.debounce(() => writeAll(store.getState()), 500, {
  */
 const startListener = () => {
   emitter.on(`CREATE_PAGE`, () => {
+    reporter.pendingActivity({ id: `requires-writer` })
     debouncedWriteAll()
   })
 
   emitter.on(`CREATE_PAGE_END`, () => {
+    reporter.pendingActivity({ id: `requires-writer` })
     debouncedWriteAll()
   })
 
   emitter.on(`DELETE_PAGE`, () => {
+    reporter.pendingActivity({ id: `requires-writer` })
     debouncedWriteAll()
   })
 
   emitter.on(`DELETE_PAGE_BY_PATH`, () => {
+    reporter.pendingActivity({ id: `requires-writer` })
     debouncedWriteAll()
   })
 }