+++ www/src/components/api-reference/index.js
@@ -1,13 +1,13 @@
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 import React from "react"
 import { graphql } from "gatsby"
 
 import DocBlock from "./doc-block"
 
-import { space } from "../../utils/presets"
-
 export default ({
   docs,
-  githubPath = null,
+  relativeFilePath = null,
   showTopLevelSignatures = false,
   ignoreParams = [],
 }) => (
@@ -16,12 +16,12 @@ export default ({
       <div
         id={definition.name}
         key={`reference list ${definition.name}`}
-        css={{ marginBottom: space[6] }}
+        sx={{ mb: 6 }}
       >
         {i !== 0 && <hr />}
         <DocBlock
           definition={definition}
-          githubPath={githubPath}
+          relativeFilePath={relativeFilePath}
           showSignature={showTopLevelSignatures}
           level={0}
           linkableTitle={true}