+++ www/src/components/tags-section.js
@@ -1,10 +1,10 @@
-import React from "react"
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 import { Link } from "gatsby"
 import TagsIcon from "react-icons/lib/ti/tags"
 const { kebabCase } = require(`lodash-es`)
 
 import Button from "./button"
-import { colors, space, fontSizes } from "../utils/presets"
 
 const TagsSection = ({ tags }) => {
   if (!tags) return null
@@ -19,32 +19,32 @@ const TagsSection = ({ tags }) => {
   })
   return (
     <div
-      css={{
+      sx={{
         display: `flex`,
         flexFlow: `row wrap`,
         justifyContent: `space-between`,
         alignItems: `baseline`,
-        borderTop: `1px solid ${colors.ui.border.subtle}`,
-        marginTop: space[10],
+        borderTop: t => `1px solid ${t.colors.ui.border}`,
+        mt: 10,
       }}
     >
       <em
-        css={{
-          fontSize: fontSizes[1],
+        sx={{
+          fontSize: 1,
           display: `block`,
           flexBasis: `60%`,
           flexGrow: 1,
           fontStyle: `normal`,
-          marginBottom: space[5],
-          marginRight: space[9],
-          marginTop: space[8],
+          mb: 5,
+          mr: 9,
+          mt: 8,
         }}
       >
         Tagged with {tagLinks}
       </em>
       <Button
         css={{ flexShrink: 0 }}
-        small
+        variant="small"
         key="blog-post-view-all-tags-button"
         to="/blog/tags"
       >